+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@dxsixpc/utils",
-  "version": "0.1.3",
+  "version": "0.2.0",
   "description": "工具库",
   "homepage": "https://dxsixpc.github.io/utils",
   "repository": "https://github.com/dxsixpc/utils",
