+++ src/helpers/index.ts
@@ -1,3 +1,2 @@
 // helpers里通常是一些较复杂的功能或函数或封装或有状态的函数
-export * from '../utils/crypto';
-export * from './request';
+export * from './request/request';
