+++ src/utils/crypto/index.ts
@@ -1,3 +1,3 @@
-export * from './CryptoAES';
-export * from './CryptoBase64';
+export { default as CryptoAES } from './CryptoAES';
+export { default as CryptoBase64 } from './CryptoBase64';
 export * from './type';
