+++ src/utils/htmlToString/index.ts
@@ -1,8 +1,1 @@
-/**
- * @name 转义的html数据还原成string
- */
-export const htmlToString = (htmlString: string) => {
-  const divDom = document.createElement('div');
-  divDom.innerHTML = htmlString;
-  return divDom.innerText || divDom.textContent || htmlString;
-};
+export { default as htmlToString } from './htmlToString';
