+++ src/utils/index.ts
@@ -1,6 +1,7 @@
 // utils文件夹里，是一些简单的无状态的静态工具函数；
+export * from './crypto';
 export * from './dataOrString';
-export * from './htmlToString';
-export * from './relativeOrAbsolute';
+export * from './htmlToString/htmlToString';
+export * from './relativeOrAbsolute/relativeToAbsolute';
 export * from './rgbaOrHex';
-export * from './styledToString';
+export * from './styledToString/styledToString';
