+++ src/utils/styledToString/type.ts
@@ -1,0 +1,3 @@
+export interface StyledType {
+  [key: string]: React.CSSProperties | StyledType | string | number;
+}
